arm: arch_timer: Allow mmio and cp15 timers to coexist
We need to support both instances of the architected timers at
runtime so that we can register an mmio timer as the broadcast
device. We also need to mark the cp15 timers as FEAT_C3_STOP
because they lost their state in certain low power modes.

Get rid of the pointer indirection and just have multiple
functions for the mmio and cp15 cases. We also assume that the
first frame is always available to us to simplify the logic. This
should be good enough for us for now.

Change-Id: I96f529744f9fa47c04efaf716e475267bae7043c
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>